refactor: 위치가 잘못된 클래스들의 패키지 변경
